Modified priceAtPurchase datatype Doble into BigDecimal

+++ ecommerce-backend/src/main/java/com/ecommerce/backend/entity/OrderItem.java
@@ -1,5 +1,7 @@
 package com.ecommerce.backend.entity;
 
+import java.math.BigDecimal;
+
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.FetchType;
@@ -28,7 +30,7 @@ public class OrderItem {
     private Integer quantity;
 
     @Column(name="price_at_purchase",nullable = false,precision = 10, scale = 2)
-    private Double priceAtPurchase;
+    private BigDecimal priceAtPurchase;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name="product_id",nullable = false)
